requirements.txt
flask>=3.0.0
feedparser>=6.0.11
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
requests>=2.32.0
lxml>=5.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=22.0.0
